feat: move the header into it's own component and updated home buttons' centering
